+++ CHANGELOG.md
@@ -719,3 +719,4 @@
 - [2026-04-26 16:21] perf(perf): optimize memory consumption in perf list viewport
 - [2026-04-26 16:31] perf(test): optimize memory consumption in test list viewport
 - [2026-04-26 17:32] test(i18n): add unit tests for i18n business logic
+- [2026-04-26 17:40] style(ui): harmonize light and dark theme palette for ui